requirements.txt
# Web application core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0

# ASR Provider dependencies
google-cloud-speech==2.21.0
elevenlabs==0.2.26
openai==1.3.0

# Analytics and reporting
elasticsearch==8.11.0
python-jose[cryptography]==3.3.0